reorder files and folders
