+++ auth/local/local.controller.js
@@ -1,9 +1,37 @@
-const { findOneUser } = require('../../../users/users.services');
-const { signToken } = require('../auth.controllers');
+const {
+  findUserByEmail,
+  findOneUser,
+} = require('../../api/users/users.services');
+
+const { signToken } = require('../auth.services');
 
 // async function changePasswordHandler(req, res) {}
 
 // async function forgotPasswordHandler(req, res) {}
+
+async function loginUserHandler(req, res) {
+  const { email, password } = req.body;
+
+  try {
+    const user = await findUserByEmail(email);
+
+    if (!user) {
+      return res.status(404).json({ message: 'User not found' });
+    }
+
+    const isMatch = await user.comparePassword(password);
+
+    if (!isMatch) {
+      return res.status(401).json({ message: 'Password does not match' });
+    }
+
+    const token = await signToken({ email: user.email });
+
+    return res.json({ token, profile: user.profile });
+  } catch (e) {
+    return res.status(500).json(e);
+  }
+}
 
 async function veryfyAccountHandler(req, res) {
   const { token } = req.params;
@@ -42,4 +70,5 @@ module.exports = {
   // changePasswordHandler,
   // forgotPasswordHandler,
   veryfyAccountHandler,
+  loginUserHandler,
 };
