+++ auth/local/local.routes.js
@@ -7,9 +7,11 @@ const {
   veryfyAccountHandler,
 } = require('./local.controller');
 
+const { validateLogin } = require('../../api/users/users.joiSchema');
+
 const router = Router();
 
-router.post('/login', loginUserHandler);
+router.post('/login', validateLogin, loginUserHandler);
 // router.post('/change-password', changePasswordHandler);
 // router.post('/forgot-password', forgotPasswordHandler);
 router.get('/ActivateAccount/:token', veryfyAccountHandler);
